Extract material anim import to own file
