+++ io_scene_nif/modules/animation/object_import.py
@@ -40,10 +40,7 @@
 import bpy
 from pyffi.formats.nif import NifFormat
 
-from io_scene_nif.modules.animation import animation_export
-from io_scene_nif.modules.object.block_registry import block_store
 from io_scene_nif.utility import nif_utils
-from io_scene_nif.utility.util_global import NifOp
 from io_scene_nif.utility.util_logging import NifLog
